requirements.txt
python=3.9
Brotli==1.0.9
click==8.0.1
colorama==0.4.4
dash==2.0.0
dash-bootstrap-components
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
et-xmlfile==1.1.0
Flask==2.0.1
Flask-Caching==1.10.1
Flask-Compress==1.10.1
future==0.18.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
openpyxl==3.0.9
pandas==1.3.3
plotly==5.3.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
tenacity==8.0.1
Werkzeug==2.0.1
gunicorn
flask
